+++ pkgs/multistockfish/lib/multistockfish.dart
@@ -1,3 +1,4 @@
 export 'src/stockfish.dart';
+export 'src/stockfish_diagnostics.dart';
 export 'src/stockfish_state.dart';
 export 'src/stockfish_flavor.dart';